WIP: Update to latest wgpu+rend3, fix errors.
This removes the hack where the tonemapping phase rendered at half
resolution. We need to replace this with a proper solution which uses
offscreen rendering next.
