+++ src/graph/graph_compiler/outputs_cache.rs
@@ -1,4 +1,4 @@
-use egui::any::TypeMap;
+use type_map::TypeMap;
 
 use crate::{prelude::graph::*, prelude::*};
 use crate::graph::poly_asm::MemAddr;
@@ -18,7 +18,7 @@ impl OutputsCache {
         addr: MemAddr<T>,
     ) {
         let cache: &mut HashMap<OutputId, MemAddr<T>> =
-            self.inner.get_mut_or_insert_with(|| Default::default());
+            self.inner.entry().or_insert_with(|| Default::default());
         cache.insert(param_id, addr);
     }
 
@@ -27,7 +27,7 @@ impl OutputsCache {
         param_id: OutputId,
     ) -> Option<MemAddr<T>> {
         let cache: &mut HashMap<OutputId, MemAddr<T>> =
-            self.inner.get_mut_or_insert_with(|| Default::default());
+            self.inner.entry().or_insert_with(|| Default::default());
         cache.get(&param_id).map(|x| *x)
     }
 }
