+++ src/graph/graph_editor_egui/editor_state.rs
@@ -1,5 +1,3 @@
-use egui::any::AnyMap;
-
 use crate::prelude::graph::*;
 use crate::prelude::*;
