+++ src/graph/graph_editor_egui/graph_node_ui.rs
@@ -1,7 +1,6 @@
 use crate::color_hex_utils::*;
 use crate::prelude::graph::*;
 use crate::prelude::*;
-use egui::any::AnyMap;
 use egui::*;
 use epaint::*;
 use glam::Vec3;
