+++ src/prelude.rs
@@ -1,15 +1,20 @@
-pub use anyhow::{Context, Result, bail, anyhow};
+pub use anyhow::{anyhow, bail, Context, Result};
 
 pub use crate::color_hex_utils::color_from_hex;
 pub use glam::{Mat4, Quat, UVec2, UVec3, Vec2, Vec3, Vec4};
 
 pub mod r3 {
     pub use rend3::{
-        types::{MaterialHandle, Mesh, MeshBuilder, MeshHandle, Object, ObjectHandle, DirectionalLight},
-        Renderer,
+        types::{
+            DirectionalLight, Handedness, MaterialHandle, Mesh, MeshBuilder, MeshHandle, Object,
+            ObjectHandle, SampleCount,
+        },
+        RenderGraph, Renderer, ReadyData,
     };
 
-    pub use rend3_routine::material::{AlbedoComponent, PbrMaterial};
+    pub use rend3_routine::base::{BaseRenderGraph, BaseRenderGraphIntermediateState};
+    pub use rend3_routine::pbr::{AlbedoComponent, PbrMaterial, PbrRoutine};
+    pub use rend3_routine::tonemapping::TonemappingRoutine;
 }
 
 pub use itertools::Itertools;
@@ -21,10 +26,9 @@ pub mod graph {
     pub use crate::graph::graph_types::*;
 }
 
-
 pub use crate::mesh::halfedge::*;
 
-pub use crate::mesh::halfedge;
 pub use crate::mesh::debug_viz;
+pub use crate::mesh::halfedge;
 
-pub use crate::math::{Vec3Ord, ToOrd, ToVec};
+pub use crate::math::{ToOrd, ToVec, Vec3Ord};
