+++ src/render_context.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use crate::{prelude::*, rendergraph::GridRoutine};
+use crate::{prelude::*, rendergraph::grid_routine::GridRoutine};
 
 use glam::Mat4;
 use rend3::{
@@ -8,23 +8,29 @@ use rend3::{
     Renderer,
 };
 use rend3_egui::EguiRenderRoutine;
-use rend3_routine::{PbrRenderRoutine, TonemappingRoutine};
+use rend3_routine::pbr::PbrRoutine;
 use wgpu::{Features, Surface, TextureFormat};
 
-use crate::rendergraph::{self, wireframe_pass::WireframeRoutine};
+use crate::rendergraph;
 
 pub struct RenderContext {
     pub renderer: Arc<Renderer>,
-    pub pbr_routine: PbrRenderRoutine,
-    pub tonemapping_routine: TonemappingRoutine,
-    pub wireframe_routine: WireframeRoutine,
+
+    pub base_graph: r3::BaseRenderGraph,
+    pub pbr_routine: r3::PbrRoutine,
+    pub tonemapping_routine: r3::TonemappingRoutine,
     pub egui_routine: EguiRenderRoutine,
     pub grid_routine: GridRoutine,
+
     pub surface: Arc<Surface>,
     pub texture_format: TextureFormat,
 
     pub objects: Vec<ResourceHandle<Object>>,
     lights: Vec<ResourceHandle<DirectionalLight>>,
+}
+
+fn ambient_light() -> Vec4 {
+    Vec4::ONE * 0.25
 }
 
 impl RenderContext {
@@ -49,23 +55,19 @@ impl RenderContext {
             rend3::types::PresentMode::Mailbox,
         );
 
-        let renderer = rend3::Renderer::new(
+        let renderer = r3::Renderer::new(
             iad,
+            r3::Handedness::Left,
             Some(window_size.width as f32 / window_size.height as f32),
         )
         .unwrap();
 
-        let render_texture_options = rend3_routine::RenderTextureOptions {
-            resolution: glam::UVec2::new(window_size.width, window_size.height),
-            samples: SampleCount::One,
-        };
-        let mut pbr_routine =
-            rend3_routine::PbrRenderRoutine::new(&renderer, render_texture_options);
-        let tonemapping_routine = rend3_routine::TonemappingRoutine::new(
-            &renderer,
-            render_texture_options.resolution,
-            format,
-        );
+        let base_graph = r3::BaseRenderGraph::new(&renderer);
+        let mut data_core = renderer.data_core.lock();
+        let pbr_routine = PbrRoutine::new(&renderer, &mut data_core, &base_graph.interfaces);
+        let tonemapping_routine =
+            r3::TonemappingRoutine::new(&renderer, &base_graph.interfaces, format);
+        drop(data_core); // Release the lock
 
         let egui_routine = EguiRenderRoutine::new(
             &renderer,
@@ -78,16 +80,12 @@ impl RenderContext {
 
         let grid_routine = GridRoutine::new(&renderer.device);
 
-        let wireframe_routine = WireframeRoutine::new(&renderer.device, &pbr_routine);
-
-        pbr_routine.set_ambient_color(glam::Vec4::ONE * 0.25);
-
         RenderContext {
             renderer,
             pbr_routine,
+            base_graph,
             tonemapping_routine,
             egui_routine,
-            wireframe_routine,
             grid_routine,
             surface,
             texture_format: format,
@@ -102,14 +100,12 @@ impl RenderContext {
 
     pub fn add_mesh_as_object(&mut self, mesh: Mesh) {
         let mesh_handle = self.renderer.add_mesh(mesh);
-        let material = rend3_routine::material::PbrMaterial {
-            albedo: rend3_routine::material::AlbedoComponent::Value(glam::Vec4::new(
-                0.8, 0.1, 0.1, 1.0,
-            )),
-            ..rend3_routine::material::PbrMaterial::default()
+        let material = r3::PbrMaterial {
+            albedo: r3::AlbedoComponent::Value(glam::Vec4::new(0.8, 0.1, 0.1, 1.0)),
+            ..Default::default()
         };
         let material_handle = self.renderer.add_material(material);
-        let object = rend3::types::Object {
+        let object = r3::Object {
             mesh: mesh_handle,
             material: material_handle,
             transform: glam::Mat4::IDENTITY,
@@ -132,7 +128,7 @@ impl RenderContext {
     }
 
     pub fn project_point(&self, point: Vec3, screen_size: Vec2) -> Vec2 {
-        let camera_manager = self.renderer.camera_manager.read();
+        let camera_manager = &self.renderer.data_core.lock().camera_manager;
 
         let clip = camera_manager.view_proj().project_point3(point);
         let clip = Vec2::new(clip.x, -clip.y);
@@ -145,7 +141,7 @@ impl RenderContext {
         self.lights.push(handle);
     }
 
-    pub fn render_frame(&mut self, egui_platform: Option<&mut egui_winit_platform::Platform>) {
+    pub fn render_frame(&mut self, egui_platform: Option<&mut egui_winit_platform::Platform>, resolution: UVec2) {
         let frame = rend3::util::output::OutputFrame::Surface {
             surface: Arc::clone(&self.surface),
         };
@@ -155,15 +151,18 @@ impl RenderContext {
 
         let mut graph = rend3::RenderGraph::new();
 
-        rendergraph::add_default_rendergraph(
+        //self.base_graph.add_to_graph(graph, ready, pbr, skybox, tonemapping, resolution, samples, ambient)
+
+        rendergraph::blackjack_rendergraph(
+            &self.base_graph,
             &mut graph,
             &ready,
             &self.pbr_routine,
-            None,
             &self.tonemapping_routine,
-            &self.wireframe_routine,
             &self.grid_routine,
-            rend3::types::SampleCount::One,
+            resolution,
+            r3::SampleCount::One,
+            ambient_light(),
         );
 
         if let Some(platform) = egui_platform {
@@ -190,16 +189,10 @@ impl RenderContext {
             rend3::types::PresentMode::Mailbox,
         );
 
-        self.renderer.set_aspect_ratio(width as f32 / height as f32 * 2.0);
+        self.renderer
+            .set_aspect_ratio(width as f32 / height as f32 * 2.0);
 
         let size = UVec2::new(width, height);
-        let options = rend3_routine::RenderTextureOptions {
-            resolution: size,
-            samples: SampleCount::One,
-        };
-
-        self.pbr_routine.resize(&self.renderer, options);
-        self.tonemapping_routine.resize(size);
         self.egui_routine.resize(width, height, scale_factor);
     }
 }
