+++ src/rendergraph/grid_shader.wgsl
@@ -11,7 +11,6 @@ struct FragmentOutput {
     [[location(0)]] color: vec4<f32>;
 };
 
-[[block]]
 struct GridRoutineUniform {
     view: mat4x4<f32>;
     proj: mat4x4<f32>;
